refactored frontend plugins loading
