+++ backend/api/init/__init__.py
@@ -4,13 +4,15 @@ import fastapi
 from pydantic import Field
 
 import nebula
+
+from nebula.plugins.frontend import PluginItemModel, get_frontend_plugins
 from nebula.settings.common import LanguageCode
+
 from server.context import ScopedEndpoint, server_context
 from server.dependencies import CurrentUserOptional
 from server.models import ResponseModel
 from server.request import APIRequest
 
-from .plugins import PluginItemModel, get_frontend_plugins
 from .settings import ClientSettingsModel, get_client_settings
