+++ backend/nebula/plugins/library.py
@@ -75,7 +75,7 @@ class PluginLibrary:
             if not os.path.isfile(manifest_path):
                 continue
 
-            log.debug("loading package", {manifest_path})
+            log.trace("Loading package", {manifest_path})
 
             try:
                 with open(manifest_path, "rb") as f:
